test(cdconc): multiple test classes for better overview
